requirements.txt
python-telegram-bot[job-queue]
python-telegram-bot
python-dotenv
praw